Refactor resource handling in DockerResourceProvider
- Simplified `GetResources` method using LINQ to create a read-only list of resources.
- Updated `InitialData` type in `ResourceSubscription` to `IReadOnlyList<Resource>` for better encapsulation.
- Added `FromContainer` method in `Resource` class to centralize resource creation logic from `ContainerListResponse`.
